==> 01_process_idats.Rmd <==
---
title: "01_process_idats"
author: "Puvvula"
date: "2023-06-30"
output: pdf_document
---

```{r}
library(pacman)
pacman::p_load(tidyverse, janitor, stringi, UpSetR, planet)
pheno_loc <- "~/Documents/methylation_pilot/"
idats<- "~/Documents/methylation_pilot/idat_files/"
processed<- "~/Documents/methylation_pilot/processed/"
```

#read chip and de-identified subject info
```{r}
pheno <- read_csv(paste0(pheno_loc, "meth_pheno.csv")) |>
  clean_names() |>
  mutate(gsm_id = paste(sentrix_id, sentrix_position, sep = "_"),
         batch = year(as.Date(date, format="%b-%d-%Y")))
  

unique_counts <- pheno |>
  group_by(sample_type) |>
  summarise(unique_participants = n_distinct(participant_id),
            n_chips = n_distinct(sentrix_id))
```

#visualize chip info
```{r}
# Generate the heatmap using ggplot2
cb_palette <- c("#000000", "#E69F00", "#56B4E9", "#009E73", "#F0E442", 
                "#0072B2", "#D55E00", "#CC79A7", "#999999", "#A69F00", 
                "#6AB4E9", "#109E73", "#E0E442", "#4072B2", "#B55E00", 
                "#EC79A7")

ggplot(pheno, aes(x = factor(sample_type), y = participant_id, fill = factor(sentrix_id))) +
  geom_tile(width = 0.8, height = 0.8) +
  scale_x_discrete(expand = c(0, 0))+
  scale_fill_manual(values = cb_palette)+
  labs(x = "Sample Type", y = "Subject ID", fill = "Sentry ID") +
  theme(#panel.grid.major = element_blank(), 
        #panel.grid.minor = element_blank(),
        panel.background = element_blank(), 
        axis.text.y = element_text(size=7),
        panel.spacing = unit(.08, "lines"),
        panel.border = element_rect(color = "black", fill = NA, size = 1),
        strip.background = element_rect(color = "black", size = 0.8),
        strip.text = element_blank(),
        axis.line = element_line(colour = "black"))+
  guides(fill = "none")+
  facet_grid(rows = vars(sentrix_id), scales = "free", space = "free")

ggsave("~/Documents/methylation_pilot/chip_seq.tiff", width = 4, height = 8.5,
       dpi = 300)
```

#================= Mehtylation data ======================#
```{r}
pacman::p_load(minfi, ENmix, IlluminaHumanMethylation450kmanifest, 
               IlluminaHumanMethylation450kanno.ilmn12.hg19, FlowSorted.Blood.450k, 
               sva, wateRmelon, RColorBrewer)
```

#read idats
```{r}
#select one type of sample to process
pheno_by_tiss <- pheno |> filter(sample_type == "CBMC")

#idat file names
idat_files<-  paste0(idats, pheno_by_tiss$gsm_id)

# get the 450k annotation data
ann450k <- getAnnotation(IlluminaHumanMethylation450kanno.ilmn12.hg19)
```

#process methylation data
```{r}
mSetSqFlt<- process_methylation_data(basenames = idat_files)
```

# calculate M-values and beta for statistical analysis
```{r}
mVals <- getM(mSetSqFlt)

bVals <- getBeta(mSetSqFlt)
```

#plot m and beta values
```{r}
par(mfrow=c(1,2))
densityPlot(bVals, sampGroups=pheno_by_tiss$sentrix_id, main="Beta values", 
            legend=FALSE, xlab="Beta values")
densityPlot(mVals, sampGroups=pheno_by_tiss$sentrix_id, main="M-values", 
            legend=FALSE, xlab="M values")
```


#reference based cell count estimation from rgset for cord blood
```{r}
rgSet <- read.metharray(basenames = idat_files) #those belong to cbmc

minfi.LC<- estimateCellCounts(rgSet, compositeCellType = "CordBlood", 
                              meanplot=TRUE, cellTypes = c("CD8T","CD4T", "NK","Bcell","Mono"),
                              returnAll = FALSE, meanPlot = TRUE, verbose = TRUE, processMethod="preprocessNoob") |>
  as.data.frame() |>
  rownames_to_column("gsm_id")

#updating cell counts
load(paste0(processed, "cbmc_processed.rda"))

cbmc_pheno <- cbmc_pheno[, -c(2:8)]

```

#=======================================
#planet reference based cell composition
```{r}
load(paste0(processed, "mp_processed.rda"))

data("plCellCpGsThird") #Third trimester reference data
row_indices <- match(rownames(plCellCpGsThird), rownames(bVals)) #subsetting CpGs that match reference data

#estimate cell count - Reference based method - placenta
houseman_estimates <- minfi:::projectCellType(
  
  # subset your data to cell cpgs
  filtered_bVals <- bVals[row_indices, , drop = FALSE],
  
  # input the reference cpg matrix
  plCellCpGsThird,
  
  lessThanOne = FALSE)

cell_est<- houseman_estimates|>
  as_tibble(rownames="gsm_id") |>
  rename_at(vars(2:7), ~ paste0("cell_mp_", .))
```

#reference free deconvolution 

```{r}
#remotes::install_version("RefFreeEWAS", version="2.2")
library(RefFreeEWAS)

#https://bmcbioinformatics.biomedcentral.com/articles/10.1186/s12859-019-3307-2
x<- RefFreeCellMix(Y = bVals, K=5)

ref_free_cells<- as_tibble(x$Omega, rownames="gsm_id") |>
  rename_at(vars(matches("^\\d+$")), ~ paste0("cell_fp_", .))
```

##==========================================================
# Export files

#export cell type counts
```{r}
cbmc_pheno <- right_join(minfi.LC, cbmc_pheno, by = "gsm_id")
```

#export rda objects for downstream analysis :: file = paste0(processed, "mp_processed.rda")
```{r}
save(bVals, mVals, mSetSqFlt, cbmc_pheno,  file = "~/Documents/methylation_pilot/processed/cbmc_processed.rda")
```
