==> 02_ewas.Rmd <==
---
title: "02_probe_wise_ewas"
author: "Puvvula"
date: "2023-07-05"
output: pdf_document
---
#https://nbis-workshop-epigenomics.readthedocs.io/en/latest/content/tutorials/methylationArray/Array_Tutorial.html

```{r}
library(pacman)
pacman::p_load(tidyverse, janitor, CpGassoc, future, future.apply, doParallel)
#processed<- "~/Documents/methylation_pilot/processed/mp_ref_cell_cnt/"
processed<- "~/Documents/methylation_pilot/processed/"
```

#CMBC analysis
```{r, message=FALSE, warning=FALSE}
#load one of these files: cbmc_processed.rda; fp_processed.rda; mp_processed.rda

load(paste0(processed, "cbmc_processed.rda"))
dat_chem_qn <- read_csv("~/Documents/methylation_pilot/exp_processed/chem_qn_fin.csv")

dat_pheno <- cbmc_pheno |>
  left_join(dat_chem_qn, by = "participant_id") |>
  mutate(across(where(is.character), as.factor)) |>
  filter(if_any(21:57, ~ !is.na(.))) |>
  mutate(across(21:57, ~ log10(. + 0.000001)))
```

#check sequence of gsm_id from preprocessed and pheno data
```{r}
bVals <- bVals[, match(dat_pheno$gsm_id, colnames(bVals)), drop = FALSE]
mVals <- mVals[, match(dat_pheno$gsm_id, colnames(mVals)), drop = FALSE]

setdiff(dat_pheno$gsm_id, colnames(bVals))
setdiff(dat_pheno$gsm_id, colnames(mVals))

nCpG = dim(bVals)[1]
```

# beta input dataset transforms into m values set in the function
```{r}
#cell types for CMBC: "CD8T", "CD4T", "NK", "Bcell", "Mono"
#cells for placenta: "cell_mp_Trophoblasts", "cell_mp_Stromal", "cell_mp_Hofbauer", "cell_mp_Endothelial", "cell_mp_nRBC"
ewas_loop_pll(exposures= c("x24_dcp", "x25_dcp", 
                           "bp_3",
                           "bpa", "bps",
                           "mecptp", "mehhtp", "monp", "m_bp", "m_bz_p", "m_cnp", "m_cop", "m_ep", "mi_bp",
                           "fluo2", "nap1", "nap2", "phen1", "phen4", "phen9", "phen23", "pyr1"),
              covariates= c("mom_age", "mom_race", "mom_edu", "pgtob", "bmi", "sex", "CD8T", "CD4T", "NK", "Bcell", "Mono"),
              bVals_list = list(bVals),
              bVals_names = list("cbmc"),
              dat_pheno = dat_pheno,
              outputFolder = "~/Documents/methylation_pilot/res_cell_cnt/cbmc_nov21_5cells")
```

#summary
```{r}  
summary_result <- ewas_summary("~/Documents/methylation_pilot/res_cell_cnt")
```


#for manhattan and volcano plot
```{r}
library(IlluminaHumanMethylation450kanno.ilmn12.hg19)
ann450k <- getAnnotation(IlluminaHumanMethylation450kanno.ilmn12.hg19) |> 
  as.data.frame() |>
  select(c("UCSC_RefGene_Name","UCSC_RefGene_Group","Relation_to_Island","chr","pos")) |>
  rownames_to_column(var = "cpg_labels")
```


#add annotation information to select ewas output files
```{r}
annotate_and_save(file_names = c("df_manhat_fluo2_fp.csv",
                                 "df_manhat_fluo2_mp.csv",
                                 "df_manhat_pyr1_fp.csv",
                                 "df_manhat_pyr1_mp.csv",
                                 "df_manhat_nap1_mp.csv",
                                 "df_manhat_nap1_fp.csv",
                                 "df_manhat_nap1_mom_cmbc.csv",
                                 "df_manhat_phen4_mp.csv",
                                 "df_manhat_m_cnp_mp.csv",
                                 "df_manhat_bpa_mp.csv",
                                 "df_manhat_e_pb_mp.csv"),
                  input_folder = "~/Documents/methylation_pilot/res_cell_cnt",
                  output_folder = "~/Documents/methylation_pilot/res_annotated_cell_cnt")
```
