==> 06_cpg_pheno.Rmd <==
---
title: "06_cpg_pheno"
author: "Puvvula"
date: "2023-09-28"
output: pdf_document
---

```{r}
library(pacman)
pacman::p_load(tidyverse, janitor)
processed<- "~/Documents/methylation_pilot/processed/"
```

#CMBC analysis
```{r, message=FALSE, warning=FALSE}
#load one of these files: cbmc_processed.rda; fp_processed.rda; mp_processed.rda

load(paste0(processed, "fp_processed.rda"))
rm(bVals, mSetSqFlt)

################################################################################
#create a dataset just with cpgs of interest
cpg_sites <- c("cg25307371", "cg09909381", "cg23175055", "cg20707765")
cpg_set<- mVals[cpg_sites, , drop = FALSE] |>
  as.data.frame() |>
  rownames_to_column(var = "cpg_site") |>
  t() |>
  as.data.frame() |>
  rownames_to_column(var = "participant_id") |>
  dplyr::slice(-1) |>
  dplyr::rename(
    "gsm_id" = "participant_id",
    "cg25307371" = "V1",
    "cg09909381" = "V2",
    "cg23175055" = "V3",
    "cg20707765" = "V4")
################################################################################

dat_chem_qn <- read_csv("~/Documents/methylation_pilot/exp_processed/chem_qn_fin.csv")

dat_pheno <- cbmc_pheno |>
  left_join(dat_chem_qn, by = "participant_id") |>
  mutate(across(where(is.character), as.factor)) |>
  filter(if_any(22:59, ~ !is.na(.))) |>
  mutate(across(22:59, ~ log10(. + 0.000001)))

# data with exposure, methylation levels and phenotypes
dat_fin <- left_join(cpg_set, dat_pheno, by="gsm_id") 

#remove all objects except the needed file
rm(list = setdiff(ls(), "dat_fin"))
write_csv(dat_fin, "~/Documents/methylation_pilot/cpg_with_outcomes.csv")
```

#scatter plot between combinations
```{r}
create_scatter_plots <- function(x_vars, y_vars, data) {
  plot_list <- list()
  
  for (x_var in x_vars) {
    for (y_var in y_vars) {
      plot <- ggplot(data, aes_string(x = x_var, y = y_var)) +
        geom_point(size = 2) +
        ylab("log - 1-hydroxynapthalene")+
        stat_smooth(method = "lm",
                    se = TRUE,  colour = "gray50") +
        theme_minimal()
      
      plot_list[[paste(x_var, y_var, sep = "_")]] <- plot
    }
  }

  grid.arrange(grobs = plot_list, nrow = length(x_vars))
}
```

```{r}
dat_fin<- read_csv("~/Documents/methylation_pilot/cpg_with_outcomes.csv") |>
  mutate(across(2:5, ~ as.numeric(.)))

create_scatter_plots(x_vars = c("cg25307371", "cg09909381", "cg23175055", "cg20707765"), 
                     y_vars = c("nap1"), 
                     data = dat_fin)
```

```{r}
x<- lm_func(dependent_vars = c("cg25307371", "cg09909381", "cg23175055", "cg20707765"),
            independent_vars= c("nap1"),
            covariates = c("bmi", "mom_age", "pgtob", "mom_edu", "mom_race", "totpg",
                           "sex", "cell_fp_1", "cell_fp_2", "cell_fp_3", "cell_fp_4", "cell_fp_5"),
            data= dat_fin,
            include_sex = FALSE,
            include_cohort = FALSE)
```
