==> 08_chemical_cell_prop_lm.Rmd <==
---
title: "08_misc_paper"
author: "Puvvula"
date: "2023-10-11"
output: pdf_document
---

```{r}
library(pacman)
p_load(tidyverse, janitor)
```

```{r}
load("~/Documents/methylation_pilot/processed/mp_ref_cell_cnt/mp_processed.rda")
load("~/Documents/methylation_pilot/processed/fp_ref_cell_cnt/fp_processed.rda")
load("~/Documents/methylation_pilot/processed/cbmc_processed.rda")
rm(bVals, mSetSqFlt, mVals)
```

#creating cell deconvolution dataframe
```{r}
cbmc_pheno<- cbmc_pheno |> select("participant_id", "CD8T", "CD4T", "NK", "Bcell", "Mono") |>
  dplyr::rename_at(vars(2:6), ~ paste0("cbmc_", .))

fp_pheno <- fp_pheno |> select("participant_id", "cell_fp_Trophoblasts", "cell_fp_Stromal", "cell_fp_Hofbauer", "cell_fp_Endothelial",
                               "cell_fp_nRBC", "cell_fp_Syncytiotrophoblast")|>
  rename_at(vars(2:7), ~ sub("^cell_", "", .))

mp_pheno<- mp_pheno |> select("participant_id", "cell_mp_Trophoblasts", "cell_mp_Stromal","cell_mp_Hofbauer",
                              "cell_mp_Endothelial", "cell_mp_nRBC", "cell_mp_Syncytiotrophoblast") |>
  rename_at(vars(2:7), ~ sub("^cell_", "", .))


cell_data <- mp_pheno |>
  full_join(fp_pheno, by = "participant_id") |>
  full_join(cbmc_pheno, by = "participant_id")
```

#box plot cell proportions
```{r}
cell_data |>
  gather(variable, value, 2:18) |>
  ggplot(aes(x = variable, y = value)) +
  geom_boxplot() +
  labs(y = "Cell proportion", x = "Cell type", title = "")+
  ylim(-0.1, 0.8)+
  theme_minimal()+
  theme(axis.text = element_text(size = 10), 
        axis.title = element_text(size = 10, face = "bold"),
        strip.text = element_text(size = 10),
        axis.text.x = element_text(angle = 35, hjust = 1, vjust = 1, size = 10),
        axis.text.y = element_text(size = 10), 
        panel.spacing.x = unit(0.02, "lines"),
        panel.spacing.y = unit(0.02, "lines"),
        strip.text.y = element_blank(),
        legend.position = "bottom", legend.box = "horizontal",
        legend.key.width = unit(0.8, "cm"),
        legend.key.height = unit(0.5, "cm")) 
```

```{r}
ggsave("~/Documents/methylation_pilot/cell_prop_boxplt.tiff",
       width = 9, height = 6, dpi=300)
```


#load exposure data and merge with cell counts
#run lines 21-36 to get cell data
```{r}
#optional analysis - requested during peer review
#converting cell proportions to log scale
cell_data_lg<- cell_data |>
  mutate(across(2:18, ~ log10(. + 0.000001)))
```


```{r}
dat_chem_qn <- read_csv("~/Documents/methylation_pilot/exp_processed/chem_qn_fin.csv")

dat_pheno <- cell_data |>
  left_join(dat_chem_qn, by = "participant_id") |>
  mutate(across(where(is.character), as.factor)) |>
  filter(if_any(22:59, ~ !is.na(.))) |>
  mutate(across(22:59, ~ log10(. + 0.000001)))
```

#clear environment
```{r}
rm(list = setdiff(ls(), "dat_pheno"))
```

#linear model between cell composition and chemical exposure
```{r}
res_df <- appliedepi::lm_func(dependent_vars = c("mp_Trophoblasts", "mp_Stromal", "mp_Hofbauer",
                                      "mp_Endothelial", "mp_nRBC", "mp_Syncytiotrophoblast",
                                      "fp_Trophoblasts", "fp_Stromal", "fp_Hofbauer",
                                      "fp_Endothelial", "fp_nRBC", "fp_Syncytiotrophoblast",
                                      "cbmc_CD8T", "cbmc_CD4T", "cbmc_NK", "cbmc_Bcell",
                                      "cbmc_Mono"),
                   independent_vars = c("x24_dcp", "x25_dcp", 
                                        "bp_3", "bpa", "bpf", "bps",
                                        "e_pb", "m_pb", "p_pb", 
                                        "mecptp", "mehhtp", "monp", "m_bp", "m_bz_p", 
                                        "m_cnp", "m_cop", "m_ep", "mi_bp",
                                        "fluo2", "nap1", "nap2",
                                        "phen1", "phen4", "phen9", "phen23", "pyr1"),
                   covariates = c("mom_age", "mom_race", "mom_edu", "pgtob", "bmi", "sex"),
                   data = dat_pheno,
                   include_sex = FALSE,
                   include_cohort = FALSE,
                   conf_level = 0.95)
```

#relabelling
```{r}
# Define the mapping for independent_variable
independent_variable_mapping <- c(
  `2,4-dichlorophenol` = "x24_dcp",
  `2,5-dichlorophenol` = "x25_dcp",
  `Benzophenone-3` = "bp_3",
  `Bisphenol-A` = "bpa",
  `Bisphenol-F` = "bpf",
  `Bisphenol-S` = "bps",
  `Ethyl Paraben` = "e_pb",
  `Methyl Paraben` = "m_pb",
  `Propyl Paraben` = "p_pb",
  `Mono-2-ethyl-5-carboxypentyl terephthalate` = "mecptp",
  `Mono-2-ethyl-5-hydrohexyl terephthalate` = "mehhtp",
  `Monooxononyl phthalate` = "monp",
  `Mono-n-butyl phthalate` = "m_bp",
  `Monobenzyl phthalate` = "m_bz_p",
  `Mono carboxyisononyl phthalate` = "m_cnp",
  `Mono carboxyisooctyl phthalate` = "m_cop",
  `Monoethyl phthalate` = "m_ep",
  `Mono-isobutyl phthalate` = "mi_bp",
  `2-Hydroxyfluorene` = "fluo2",
  `1-Hydroxynaphthalene` = "nap1",  # Corrected typo in variable name
  `2-Hydroxynaphthalene` = "nap2",  # Corrected typo in variable name
  `1-Hydroxyphenanthrene` = "phen1",
  `2,3-Hydroxyphenanthrene` = "phen23",
  `4-Hydroxyphenanthrene` = "phen4",
  `9-Hydroxyphenanthrene` = "phen9",
  `1-Hydroxypyrene` = "pyr1"
)

res_df_fin <- res_df |>
  filter(cohort_level == "all" & sex_level == "all") |>
  mutate(across(all_of(c("coefficient", "ci_lower", "ci_upper")), ~ sprintf("%.2f (%.2f, %.2f)", .x, ci_lower, ci_upper))) |>
  mutate(independent_variable = fct_recode(independent_variable, !!!independent_variable_mapping))  |>
  select(-c(ci_lower, ci_upper, sex_level, cohort_level)) |>
  mutate(dna_m_smpl = if_else(grepl("^cbmc_", dependent_variable), "CBMC",
                                          if_else(grepl("^fp_", dependent_variable), "Fetal placenta",
                                                  if_else(grepl("^mp_", dependent_variable), "Maternal placenta", dependent_variable) ) ) ) |>
  mutate(dependent_variable = str_replace(dependent_variable, "^(mp_|cbmc_|fp_)", ""))
  
#Save results
write_csv(res_df_fin, "~/Documents/methylation_pilot/chem_cell_prop_lm.csv")
```


#Prep res_df_fin table for heat map
```{r}
res_df_fin<- read_csv("~/Documents/methylation_pilot/chem_cell_prop_lm.csv")

vis_df <- res_df_fin |>
  separate(coefficient, into = c("coef", "lci", "uci"), sep = "\\s+\\(|,|\\)") |>
  mutate(
    across(c(coef, lci, uci), as.numeric),
    contains_zero = ifelse(lci > 0 | uci < 0, FALSE, TRUE),
    dependent_variable = fct_recode(dependent_variable,
                                    "NK cells" = "NK",
                                    "Monocytes" = "Mono"),
    independent_variable = fct_recode(independent_variable,
                                      "Mono-2-ethyl-5-hydroxyhexyl terephthalate" = "Mono-2-ethyl-5-hydrohexyl terephthalate"),
    independent_variable = factor(independent_variable, levels = rev(c(
      "1-Hydroxynaphthalene", "2-Hydroxynaphthalene", "2-Hydroxyfluorene", "1-Hydroxyphenanthrene", 
      "2,3-Hydroxyphenanthrene", "4-Hydroxyphenanthrene","9-Hydroxyphenanthrene", "1-Hydroxypyrene", 
      "Monoethyl phthalate", "Mono-isobutyl phthalate", "Mono carboxyisooctyl phthalate",
      "Mono carboxyisononyl phthalate", "Monobenzyl phthalate", "Mono-n-butyl phthalate",
      "Monooxononyl phthalate",
      "Mono-2-ethyl-5-hydroxyhexyl terephthalate",
      "Mono-2-ethyl-5-carboxypentyl terephthalate",
      "2,4-dichlorophenol", "2,5-dichlorophenol", 
      "Benzophenone-3", "Bisphenol-A", "Bisphenol-F", "Bisphenol-S", 
      "Methyl Paraben", "Ethyl Paraben", "Propyl Paraben")
    ))
  ) |>
  filter(!(independent_variable %in% c("Methyl Paraben", "Ethyl Paraben", "Propyl Paraben", "Bisphenol-F")))

# Reorder the levels within each facet of dna_m_smpl
vis_df <- vis_df |>
  group_by(dna_m_smpl) |>
  mutate(dependent_variable = fct_relevel(dependent_variable, c("CD4T", "CD8T", "Bcell",
                                                               "Monocytes", "NK cells",
                                                               "Trophoblasts","Syncytiotrophoblast", 
                                                               "Stromal", "Hofbauer", "Endothelial"))) |>
  ungroup()
```

#visualize lm results between chemical exposure and cell composition - export
```{r}
vis_df |>
  ggplot(aes(y = independent_variable, x = dependent_variable, 
             fill = coef, label=contains_zero)) +
  geom_tile() +
  geom_text(aes(label = ifelse(contains_zero, "", "x"), size = 5), show.legend = FALSE) +
  facet_grid(. ~ dna_m_smpl, scales = "free_x", space = "free_x")+
  scale_fill_gradientn(
    colors = c("lightpink", "white", "lightblue"),
    values = scales::rescale(c(min(vis_df$coef, na.rm = TRUE), 0, max(vis_df$coef, na.rm = TRUE))),
    breaks = c(min(vis_df$coef, na.rm = TRUE), 0, max(vis_df$coef, na.rm = TRUE)),
    labels = c(min(vis_df$coef, na.rm = TRUE), 0, max(vis_df$coef, na.rm = TRUE)),
    guide = guide_colorbar(
      title = expression(beta[italic(lm)]),
      title.position = "left",
      title.hjust = 1,
      title.vjust = 1
    )
  )  +
  theme_bw() +
  labs(x = "", y = "", fill = "coefficient") +
  theme(axis.text = element_text(size = 10), 
        axis.title = element_text(size = 10, face = "bold"),
        strip.text = element_text(size = 10),
        panel.grid.minor = element_blank(),
        panel.grid.major = element_blank(),
        panel.grid.major.y = element_blank(),
        axis.text.x = element_text(angle = 35, hjust = 1, vjust = 1, size = 10),
        axis.text.y = element_text(size = 10), 
        panel.spacing.x = unit(0.02, "lines"),
        panel.spacing.y = unit(0.02, "lines"),
        strip.text.y = element_blank(),
        legend.position = "bottom", legend.box = "horizontal",
        legend.key.width = unit(0.8, "cm"),
        legend.key.height = unit(0.5, "cm")) 

ggsave("~/Documents/methylation_pilot/chem_cell_prop_lm_viz_out_lg.tiff",
       width = 8, height = 7, dpi=300)
```
