==> 09_dmr.Rmd <==
---
title: "03_dmr_analysis"
author: "Puvvula"
date: "2023-08-01"
output: html_document
---

```{r}
library(pacman)
pacman::p_load(tidyverse, janitor)
processed<- "~/Documents/methylation_pilot/processed/"
```

#CMBC analysis
```{r}
load(paste0(processed, "cbmc_processed.rda"))
dat_chem_qn <- read_csv("~/Documents/methylation_pilot/exp_processed/chem_qn_fin.csv")

dat_pheno <- cbmc_pheno |>
  left_join(dat_chem_qn, by = "participant_id") |>
  mutate(across(where(is.character), as.factor)) |>
  filter(if_any(22:59, ~ !is.na(.))) |>
  mutate(across(22:59, ~ log10(. + 0.000001)))
```

#converting exposure variables to binary for dmr analysis
```{r}
dat_pheno<- dat_pheno |>
  mutate(across(22:59, ~if_else(. < median(.), "LT_med", "GT_med")))
```

#design matrix with binary exposure
```{r}
design <- model.matrix(~bpa+CD8T+CD4T+NK+Bcell+Mono+Gran+nRBC, data=dat_pheno)
```

```{r}
ann450k = getAnnotation(IlluminaHumanMethylation450kanno.ilmn12.hg19)
ann450kSub <- ann450k[match(rownames(mVals),ann450k$Name),
                      c(1:4,12:19,24:ncol(ann450k))]
```

```{r}
suppressMessages(library(DMRcate))
```

```{r}
myAnnotation <- cpg.annotate(object = bVals,
                             datatype = "array",
                             what = "Beta",
                             analysis.type = "differential",
                             design = design,
                             arraytype = "450K", 
                             coef = 2)
```


```{r}
DMRs <- dmrcate(myAnnotation, lambda=1000, C=2)
DMRs
# Create GRanges object; create directory when prompted
results.ranges <- extractRanges(DMRs)
results.ranges
```


```{r}
# set up the grouping variables and colours
dat_pheno$nap1<- as.factor(dat_pheno$nap1)
cols= c("magenta", "red")[dat_pheno$nap1]
names(cols) = levels(dat_pheno$nap1)[dat_pheno$nap1]

# draw the plot for the second DMR - first gives error for some reason...
DMR.plot(ranges = results.ranges,
         dmr = 2,
         CpGs = mSetSqFlt,
         phen.col = cols,
         genome = "hg19")
```
